--- src/components/avatar/userAvatar.tsx
@@ -1,21 +1,0 @@
-import { Avatar } from '@chakra-ui/react';
-
-export type UserAvatarProps = {
-    firstName: string;
-    lastName: string;
-};
-
-const UserAvatar = (props: UserAvatarProps) => {
-    return (
-        <>
-            <Avatar
-                size='md'
-                name={props.firstName + ' ' + props.lastName}
-                src={`https://ui-avatars.com/api/?name=${
-                    props.firstName + '+' + props.lastName
-                }`}
-            />
-        </>
-    );
-};
-export default UserAvatar;
