+++ src/pages/connect/events.tsx
@@ -1,134 +1,15 @@
-import ConnectLayout from '@/components/layouts/ConnectLayout';
+import ConnectLayout from 'layouts/ConnectLayout';
 import {
     Box,
-    Button,
-    Flex,
     Heading,
-    HStack,
-    Tag,
-    Text,
-    Tooltip,
-    useToast,
     VStack,
 } from '@chakra-ui/react';
 import { supabaseClient, withPageAuth } from '@supabase/auth-helpers-nextjs';
-import { useUser } from '@supabase/auth-helpers-react';
 import { useProfile } from 'hooks/profile';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { GrowEvent, GrowEventDto, NextPageWithLayout } from 'types';
+import GrowEventCard from '@/components/events/GrowEventCard';
 
-function GrowEventCard({
-    event,
-    registered,
-}: {
-    event: GrowEvent;
-    registered: boolean;
-}) {
-    const [registeredLocal, setRegisteredLocal] = useState(registered);
-    const { user } = useUser();
-    const toast = useToast();
-    const { day, month } = useMemo(() => {
-        const day = String(event.date.getDate()).padStart(2, '0');
-        const month = event.date.toLocaleString('en-US', { month: 'short' });
-        return { day, month };
-    }, [event.date]);
-    const over = new Date() > new Date(event.date);
-
-    async function register() {
-        const { error } = await supabaseClient
-            .from('registrations')
-            .insert({ user_id: user?.id, event_id: event.id });
-        if (error) {
-            toast({
-                title: 'Something went wrong...',
-                status: 'error',
-                duration: 4000,
-                isClosable: true,
-            });
-        } else {
-            toast({
-                title: 'Registered to ' + event.title,
-                status: 'success',
-                duration: 4000,
-                isClosable: true,
-            });
-            setRegisteredLocal(true);
-        }
-    }
-
-    async function withdrawRegistration() {
-        const { error } = await supabaseClient
-            .from('registrations')
-            .delete()
-            .match({ event_id: event.id, user_id: user?.id });
-        if (error) {
-            toast({
-                title: 'Something went wrong...',
-                status: 'error',
-                duration: 4000,
-                isClosable: true,
-            });
-        } else {
-            toast({
-                title: 'Unregistered from ' + event.title,
-                status: 'success',
-                duration: 4000,
-                isClosable: true,
-            });
-            setRegisteredLocal(false);
-        }
-    }
-
-    let actionButton;
-    if (event.mandatory) {
-        actionButton = (
-            <Tooltip label="Event is mandatory" shouldWrapChildren>
-                <Button disabled={true} size="sm" variant="outline">
-                    Withdraw Registration
-                </Button>
-            </Tooltip>
-        );
-    } else if (registeredLocal) {
-        actionButton = (
-            <Button onClick={withdrawRegistration} size="sm" variant="outline">
-                Withdraw Registration
-            </Button>
-        );
-    } else {
-        actionButton = (
-            <Button
-                onClick={register}
-                color="primary"
-                size="sm"
-                variant="outline"
-            >
-                Register
-            </Button>
-        );
-    }
-
-    return (
-        <HStack
-            gap={4}
-            alignItems="start"
-            color={over ? 'gray.500' : 'inherit'}
-        >
-            <Flex justify="space-between" w="3.5rem" fontWeight="semibold">
-                <Text>{day}</Text>
-                <Text>{month}</Text>
-            </Flex>
-            <VStack alignItems="start">
-                <HStack>
-                    <Heading size="md">{event.title}</Heading>
-                    {event.mandatory ? <Tag>mandatory</Tag> : undefined}
-                    {event.online ? <Tag>online</Tag> : undefined}
-                </HStack>
-                <Text>{event.description}</Text>
-                {!over ? actionButton : undefined}
-            </VStack>
-        </HStack>
-    );
-}
 
 const EventsPage: NextPageWithLayout = () => {
     const { profile } = useProfile();
